hotfix for final push in git/helpers
This is just a hotfix as the real problem remains.
Main documentation is in #98, but the gist of it
is that I worked around the problem by not using
git push --recurse-submodules=on-demand when pushing
in the root collection after a commit in a submodule
as this somehow throws an error looking like this:

> git push --recurse-submodules=on-demand
Locking support detected on remote "books". Consider enabling it with:
  $ git config lfs.$submodule-git-url.git/info/lfs.locksverify true
fatal: remote 'books' not configured
fatal: process for submodule 'stories' failed
fatal: the remote end hung up unexpectedly
Gitea: Internal error

No idea why, hotfix should be enough for now

Fixes #98